Redesigned websocket message marshalling/unmarshalling
WebSocketMessages will now store the payload as interface{}
and be automatically marshalled and unmarshalled. This requires
prior registering of the necessary types and typeIDs with the use
of the RegisterMessageType function. The unmarshalling process uses
reflection and the type stored in the type map to create a payload
instance of the appropriate type.