Create an inital version of a web frontend based on SB admin 2.0
The frontend uses the SB Admin Angular 1.2 template. To run you
will need to have bower installed and run bower install in the
master/views directory to install all the dependencies. It has
working basic authentication, can log in, log out and relogin if
a session cookie is present. Angular dependency was updated to 1.5.x.